@@ -18,16 +18,16 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.1.0 <4.0.0"
 
 dependencies:
+  cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
   flutter_launcher_icons: ^0.9.1
-  google_fonts: ^2.1.0
-  http: ^0.13.3
-  provider: ^4.3.2+3
+  google_fonts: ^5.1.0
+  http: ^1.1.0
+  provider: ^6.0.5
   url_launcher: ^6.0.9
 
 flutter_icons:
@@ -36,6 +36,7 @@ flutter_icons:
   image_path: "assets/icon_logo.png"
 
 dev_dependencies:
+  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
 
