@@ -1,8 +1,8 @@
 class City {
-  int id;
-  String name;
-  String imageUrl;
-  bool isPopular;
+  int? id;
+  String? name;
+  String? imageUrl;
+  bool? isPopular;
 
   City({this.id, this.imageUrl, this.name, this.isPopular = false});
 }
