@@ -1,8 +1,8 @@
 class Tips {
-  int id;
-  String title;
-  String imageUrl;
-  String updatedAt;
+  int? id;
+  String? title;
+  String? imageUrl;
+  String? updatedAt;
 
   Tips({
     this.id,
