@@ -25,7 +25,7 @@ class SpaceCard extends StatelessWidget {
               child: Stack(
                 children: [
                   Image.network(
-                    space.imageUrl,
+                    space.imageUrl.toString(),
                     width: 130,
                     height: 110,
                     fit: BoxFit.cover,
@@ -66,7 +66,7 @@ class SpaceCard extends StatelessWidget {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               Text(
-                space.name,
+                space.name.toString(),
                 style: blackTextStyle.copyWith(fontSize: 18),
               ),
               SizedBox(
